# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/disease": {
            "get": {
                "tags": [
                    "Disease API CRUD"
                ],
                "summary": "Get all Diseases",
                "description": "Retrieve all Diseases",
                "operationId": "disease_index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disease"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Disease API CRUD"
                ],
                "summary": "Create a new Disease",
                "description": "Add a new Disease",
                "operationId": "disease_store",
                "requestBody": {
                    "description": "Disease save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "treatment"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "treatment": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disease"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/v1/disease/{id}": {
            "get": {
                "tags": [
                    "Disease API CRUD"
                ],
                "summary": "Get a single Disease by ID",
                "description": "Retrieve a single Disease by its ID",
                "operationId": "disease_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Disease to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disease"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Disease API CRUD"
                ],
                "summary": "Update a Disease by ID",
                "description": "Update a specific Disease by its ID",
                "operationId": "disease_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Disease to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Disease save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "treatment"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "treatment": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "_method": {
                                        "description": "Read-only: This property cannot be modified.",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disease"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Disease API CRUD"
                ],
                "summary": "Delete a Disease by ID",
                "description": "Remove a specific Disease by its ID",
                "operationId": "disease_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Disease to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Disease"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Disease": {
                "title": "Disease title",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "treatment": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": []
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            }
        }
    }
}